+++ src/Pages/Validator/Components/ValidatorCommission.js
@@ -6,8 +6,11 @@ import { numberFormat, formatNhash } from 'utils';
 
 const ValidatorCommission = () => {
   const [showBondedPopup, setShowBondedPopup] = useState(false);
-  const [showCommissionPopup, setShowCommissionPopup] = useState(false);
-  const { getValidatorCommission, validatorCommission, validatorCommissionLoading } = useValidators();
+  const {
+    getValidatorCommission,
+    validatorCommission,
+    validatorCommissionLoading,
+  } = useValidators();
   const { validatorId } = useParams();
   // Get validatorCommission info on load
   useEffect(() => {
@@ -24,15 +27,23 @@ const ValidatorCommission = () => {
     commissionRewards = {},
   } = validatorCommission;
   const { count: bondedTokensCount } = bondedTokens;
-  const { maxRate: commissionMaxRate, rate: commissionRateAmount } = commissionRate;
+  const {
+    maxChangeRate: commissionMaxChangeRate,
+    maxRate: commissionMaxRate,
+    rate: commissionRateAmount,
+  } = commissionRate;
   const { amount: commissionRewardsAmount } = commissionRewards;
   const { count: delegatorBondedCount, denom: delegatorBondedDenom } = delegatorBonded;
   const { count: selfBondedCount, denom: selfBondedDenom } = selfBonded;
 
-  const delegatorBondedPercent = `${numberFormat((delegatorBondedCount / bondedTokensCount) * 100, 2)} %`;
+  const delegatorBondedPercent = `${numberFormat(
+    (delegatorBondedCount / bondedTokensCount) * 100,
+    2
+  )} %`;
   const selfBondedPercent = `${numberFormat((selfBondedCount / bondedTokensCount) * 100, 2)} %`;
   const commisionRatePercent = `${numberFormat(commissionRateAmount * 100, 2)} %`;
   const commissionRatePercent = `${numberFormat(commissionMaxRate * 100, 2)} %`;
+  const commissionMaxChangeRatePercent = `${numberFormat(commissionMaxChangeRate * 100, 2)} %`;
 
   const popupNoteBondedTokens = {
     visibility: { visible: showBondedPopup, setVisible: setShowBondedPopup },
@@ -46,15 +57,11 @@ const ValidatorCommission = () => {
       },
       {
         title: 'Delegator Bonded:',
-        value: `${formatNhash(delegatorBondedCount)} ${delegatorBondedDenom} (${delegatorBondedPercent})`,
+        value: `${formatNhash(
+          delegatorBondedCount
+        )} ${delegatorBondedDenom} (${delegatorBondedPercent})`,
       },
     ],
-  };
-  const popupNoteCommissionRateRange = {
-    visibility: { visible: showCommissionPopup, setVisible: setShowCommissionPopup },
-    icon: { name: 'HELP_OUTLINE', size: '1.7rem' },
-    method: ['click', 'hover'],
-    fontColor: 'FONT_WHITE',
   };
 
   const summaryData = [
@@ -67,11 +74,8 @@ const ValidatorCommission = () => {
     { title: 'Delegators', value: delegatorCount },
     { title: 'Total Shares', value: `${formatNhash(totalShares)} hash` },
     { title: 'Commission Rewards', value: `${formatNhash(commissionRewardsAmount)} hash` },
-    {
-      title: 'Commission Rate Range',
-      value: `0 ~ ${commissionRatePercent}`,
-      popupNote: popupNoteCommissionRateRange,
-    },
+    { title: 'Commission Rate Range', value: `0 ~ ${commissionRatePercent}` },
+    { title: 'Max Change Rate', value: commissionMaxChangeRatePercent },
   ];
 
   return (